# Micro-optimizations and rustfmt