```
running 4 tests
test bench_line_hasan_10  ... bench:         538 ns/iter (+/- 18)
test bench_line_hasan_100 ... bench:      36,139 ns/iter (+/- 696)
test bench_line_tan_10    ... bench:         740 ns/iter (+/- 20)
test bench_line_tan_100   ... bench:      40,415 ns/iter (+/- 807)
```

tests on a 10x10 and a 100x100 cell grid respectively. :ok_hand: