@expectocode 2019-02-22
Nice stuff. Could you put some of the values from the String allocation into constants? There's a lot of magic numbers now