@JuanPotato 2019-02-22
Not sure how to make it.
```(self.width + 2) * (self.height + 2) / 2 * 3```
That should be width + 3. Two for the border, one for newline.
height + 2, 2 lines for border lines. 
/ 2 because two gridlines per character line
*3 because 3 bytes per block character

How do you want to chop that up