@expectocode 2019-02-22
How about width * height * NUM_BYTES_CELL_BLOCK + width * NUM_SIDES + height * NUM_SIDES + NUM_CORNERS