@expectocode 2019-02-22
wait, are we accounting for newlines?