@expectocode 2019-02-24
@JuanPotato could you see how this impacts your benchmarks?